+++ scripts/lichessTools-init.js
@@ -125,6 +125,7 @@ lichessTools.loadTool(LiChessTools.Tools.CommentEditing);
 lichessTools.loadTool(LiChessTools.Tools.PlayerExtraButtons);
 lichessTools.loadTool(LiChessTools.Tools.CevalCache);
 lichessTools.loadTool(LiChessTools.Tools.ShowKmaps);
+lichessTools.loadTool(LiChessTools.Tools.ChallengeOptions);
 
 
 lichessTools.loadTool(LiChessTools.Tools.HideBoardCommand);